+++ android/src/main/java/com/swmansion/gesturehandler/react/RNGestureHandlerPackage.java
@@ -3,17 +3,10 @@ package com.swmansion.gesturehandler.react;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
-import com.facebook.react.common.MapBuilder;
-import com.facebook.react.uimanager.ThemedReactContext;
-import com.facebook.react.uimanager.ViewGroupManager;
 import com.facebook.react.uimanager.ViewManager;
-import com.facebook.react.views.view.ReactViewManager;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
-
-import androidx.annotation.Nullable;
 
 public class RNGestureHandlerPackage implements ReactPackage {
 
@@ -26,6 +19,7 @@ public class RNGestureHandlerPackage implements ReactPackage {
   public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
     return Arrays.<ViewManager>asList(
             new RNGestureHandlerRootViewManager(),
-            new RNGestureHandlerButtonViewManager());
+            new RNGestureHandlerButtonViewManager(),
+            new RNGestureHandlerEventReceiverViewManager());
   }
 }
