+++ examples/Example/src/prot_reanimated/index.tsx
@@ -58,7 +58,7 @@ function Draggable() {
         'worklet';
         offsetX.value = e.translationX + startX.value;
         offsetY.value = e.translationY + startY.value;
-        //console.log(a);
+        //console.log('a');
       },
     })
   );
